requirements.txt
django>=5.0,<5.2
argon2-cffi>=23.1.0
python-dotenv>=1.0.0
django-environ>=0.11.2
cryptography>=41.0.0

